issue #1015 - copied blockly with enabled insert operation and different index positions for calliope - extracted the implementation of set and get to AbstractCppVisitor from Arduino and Calliope specific

+++ OpenRobertaParent/RobotMbed/resources/libs2017/microbit/inc/NEPODefs.h
@@ -173,12 +173,12 @@ double _getListMedian(std::list<double> &list) {
     sorted.sort();
     auto iterator = sorted.begin();
     if( sorted.size() % 2 == 0 ) {
-        for( int i = 0 ; i < sorted.size() / 2 ; i ++ ) {
+        for( uint16_t i = 0 ; i < sorted.size() / 2 ; i ++ ) {
             iterator++;
         }
         return (*iterator + *--iterator ) / 2;
     } else {
-        for( int i = 0 ; i < sorted.size() / 2 ; i ++ ) {
+        for( uint16_t i = 0 ; i < sorted.size() / 2 ; i ++ ) {
             iterator++;
         }
         return *iterator;
